Option to change test notification content
Added option for changing test notification content, so it will be possible to test regex
